split out the make program
