+++ scripts/src/Program/Make.hs
@@ -1,0 +1,16 @@
+module Program.Make where
+
+import Control.Monad.IO.Class
+import GHC.Conc (getNumProcessors)
+import Program
+import Utils
+
+data Make
+instance Program Make where
+    executableName = "make"
+
+make :: (MonadIO m) => FilePath -> m ()
+make location = do
+    cpuCount <- liftIO $ getNumProcessors
+    jobCount <- getEnvDefault "JOB_COUNT" (show cpuCount)
+    callCwd @Make location ["-j", jobCount]
